Second
This is great, but I wouldd like to make a couple of changes to the HTML.  Under the traits (strength, dexterity, etc...) change "Save" to "Saving Throw" and put a small numeric text entry field in front of it.  Also, instead of "Skill Name" put the specific skills from the picture ("Athletics" "Acrobatics" etc...).  You can drop the page break between the two pictures and leave them both in a single shared element.  Expand the boxes of "Species Traits" and "Feats" to take up a little more vertical space.  Increase the size of the "Backstory and Personality" box to fill the vertical space on that page.